requirements.txt
nemo_toolkit[asr]
sounddevice
numpy
scipy
pynput
pyperclip
torch
pyinstaller
